streaming_iterator
Trait DoubleEndedStreamingIteratorMut

pub trait DoubleEndedStreamingIteratorMut: DoubleEndedStreamingIterator + StreamingIteratorMut {
    // Provided methods
    fn next_back_mut(&mut self) -> Option<&mut Self::Item> { ... }
    fn rfold_mut<B, F>(self, init: B, f: F) -> B
       where Self: Sized,
             F: FnMut(B, &mut Self::Item) -> B { ... }
}

A mutable streaming iterator able to yield elements from both ends.

Provided Methods

fn next_back_mut(&mut self) -> Option<&mut Self::Item>

Advances the iterator and returns the next mutable value from the back.

The behavior of calling this method after the end of the iterator has been reached is unspecified.

The default implementation simply calls advance_back followed by get_mut.

fn rfold_mut<B, F>(self, init: B, f: F) -> B
where Self: Sized, F: FnMut(B, &mut Self::Item) -> B,

Reduces the iterator’s mutable elements to a single, final value, starting from the back.